Updating group controller

+++ src/main/java/br/com/compass/challenge2/controller/GroupController.java
@@ -1,8 +1,12 @@
 package br.com.compass.challenge2.controller;
 
 import br.com.compass.challenge2.entity.Group;
+import br.com.compass.challenge2.entity.Student;
 import br.com.compass.challenge2.service.GroupService;
+import jakarta.validation.Valid;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
@@ -19,53 +23,71 @@ public class GroupController {
     }
 
     @GetMapping("/all")
-    public List<Group> findAll(){
-        return groupService.findAll();
+    public ResponseEntity<List<Group>> findAll(){
+        List<Group> groups = groupService.findAll();
+        return new ResponseEntity<>(groups, HttpStatus.OK);
     }
 
     @GetMapping("/{groupId}")
-    public Group findGroupById(@PathVariable Long groupId){
+    public ResponseEntity<Group> findGroupById(@PathVariable Long groupId){
         Group group = groupService.findById(groupId);
         if(group == null){
-            throw new RuntimeException("Group not found! Id: "+groupId);
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
         }
-        return group;
+        return new ResponseEntity<>(group, HttpStatus.OK);
     }
 
     @GetMapping("/{groupName}")
-    public Group findGroupByName(@PathVariable String groupName){
+    public ResponseEntity<Group> findGroupByName(@PathVariable String groupName){
         Group group = groupService.findDistinctByName(groupName);
         if(group == null){
-            throw new RuntimeException("Group not found! Name: "+groupName);
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
         }
-        return group;
+        return new ResponseEntity<>(group, HttpStatus.OK);
     }
 
     @PostMapping
-    public Group addGroup(@RequestBody Group postGroup){
+    public ResponseEntity<Group> createGroup(@Valid @RequestBody Group postGroup){
         postGroup.setId(0L);
         Group newGroup = groupService.save(postGroup);
-        return newGroup;
+        return new ResponseEntity<>(newGroup, HttpStatus.CREATED);
     }
 
-    @PutMapping
-    public Group updateGroup(@RequestBody Group putGroup){
-        Group newGroup = groupService.save(putGroup);
-        return newGroup;
+    @PutMapping("/{id}")
+    public ResponseEntity<Group> updateGroup(@Valid @PathVariable Long id, @RequestBody Group putGroup){
+        Group existingGroup = groupService.findById(id);
+        if(putGroup != null){
+            putGroup.setId(id);
+            Group updatedGroup = groupService.update(existingGroup);
+            return new ResponseEntity<>(updatedGroup, HttpStatus.OK);
+        }
+        else{
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
     }
 
     @DeleteMapping("/{groupId}")
-    public String deleteGroupById(@PathVariable Long groupId){
-        //Group checkGroup = groupService.findById(groupId);
-        groupService.deleteById(groupId);
-        return "Deleted employee with id: "+groupId;
+    public ResponseEntity<Void> deleteGroupById(@PathVariable Long groupId){
+        Group checkGroup = groupService.findById(groupId);
+        if(checkGroup != null){
+            groupService.deleteById(groupId);
+            return new ResponseEntity<>(HttpStatus.NO_CONTENT);
+        }
+        else{
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
     }
 
     @DeleteMapping("/{groupName}")
-    public String deleteGroupByName(@PathVariable String groupName){
-        //Group checkGroup = groupService.findDistinctByName(groupName);
-        groupService.deleteByName(groupName);
-        return "Deleted employee with name: "+groupName;
+    public ResponseEntity<Void> deleteGroupByName(@PathVariable String groupName){
+        Group checkGroup = groupService.findDistinctByName(groupName);
+        if(checkGroup != null){
+            groupService.deleteByName(groupName);
+            return new ResponseEntity<>(HttpStatus.NO_CONTENT);
+        }
+        else{
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
     }
 
 }
